requirements.txt
streamlit>=1.31
plotly>=5.20
pandas>=2.1
pyyaml>=6.0
python-dotenv>=1.0
duckdb>=0.10
chromadb>=0.5
boto3>=1.34
reportlab>=4.0
kaleido>=0.2
# Optional ag-Grid support:
streamlit-aggrid>=0.3.4
hnswlib
